requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.3.1
pymupdf>=1.23.0
pdfminer.six>=20221105
tiktoken>=0.5.0
